Workaround for missing cached tag list
When the article is created from the factory the cache tag list is not
set. Saving the article appears to populate the column for us (which
is referenced in `calculate_implicit_points` using pluck)

If cache_tag_list should not be treated as the source of truth for the
Follows::UpdatesPointsWorker then it needs to replace both the
decorated cached tag list array in `perform` and the
pluck(:cached_tag_list) in `calculate_implicit_points`.

When we _create_ the article the cached_tag_list is not set (it is on
subsequent saves). Rather than patching the factory to set this (which
is not how we do this otherwise), update the points worker to use
tag_list in update points worker, rather than cached_tag_list.

Setting the cached_tag_list column in the articles factory would also
work (see 33dd82349e309559836027f2a31ad390971cbb55).